refactor: remove unnecessary nested index.ts
