+++ app/src/pages/Editor/components/Canvas.tsx
