--- app/src/pages/Editor/components/Canvas/index.tsx
@@ -1,3 +1,0 @@
-import Canvas from './Canvas';
-
-export default Canvas;
