+++ app/src/pages/Editor/components/ContentSidebar.tsx
