--- app/src/pages/Editor/components/ContentSidebar/index.tsx
@@ -1,3 +1,0 @@
-import ContentSidebar from './ContentSidebar';
-
-export default ContentSidebar;
