--- app/src/pages/Editor/components/MainButtons/index.tsx
@@ -1,3 +1,0 @@
-import MainButtons from './MainButtons';
-
-export default MainButtons;
