--- app/src/pages/Editor/components/MenuBurger/index.tsx
@@ -1,3 +1,0 @@
-import MenuBurger from './MenuBurger';
-
-export default MenuBurger;
