+++ app/src/pages/Home/Home.tsx
@@ -1,7 +1,5 @@
-const Home = () => (
+export const Home = () => (
   <>
     Metagallery landing page
   </>
 );
-
-export default Home;
